Move Image implementations from `.message` to `.message.image`
