+++ mirai-core/src/commonMain/kotlin/contact/GroupImpl.kt
@@ -26,7 +26,13 @@ import net.mamoe.mirai.internal.QQAndroidBot
 import net.mamoe.mirai.internal.contact.announcement.AnnouncementsImpl
 import net.mamoe.mirai.internal.contact.file.RemoteFilesImpl
 import net.mamoe.mirai.internal.contact.info.MemberInfoImpl
-import net.mamoe.mirai.internal.message.*
+import net.mamoe.mirai.internal.message.MiraiInternalMessageFlag
+import net.mamoe.mirai.internal.message.OfflineAudioImpl
+import net.mamoe.mirai.internal.message.contextualBugReportException
+import net.mamoe.mirai.internal.message.image.OfflineGroupImage
+import net.mamoe.mirai.internal.message.image.calculateImageInfo
+import net.mamoe.mirai.internal.message.image.getIdByImageType
+import net.mamoe.mirai.internal.message.image.getImageTypeById
 import net.mamoe.mirai.internal.network.components.BdhSession
 import net.mamoe.mirai.internal.network.handler.logger
 import net.mamoe.mirai.internal.network.highway.ChannelKind
