+++ mirai-core/src/commonMain/kotlin/message/MultiMsgUploader.kt
@@ -12,6 +12,7 @@ package net.mamoe.mirai.internal.message
 import io.ktor.utils.io.core.*
 import net.mamoe.mirai.internal.contact.SendMessageHandler
 import net.mamoe.mirai.internal.contact.takeSingleContent
+import net.mamoe.mirai.internal.message.source.MessageSourceInternal
 import net.mamoe.mirai.internal.network.QQAndroidClient
 import net.mamoe.mirai.internal.network.highway.Highway
 import net.mamoe.mirai.internal.network.highway.ResourceKind
